Prototype scratchpad for oVirt API experiments.
Use like this: ../../nbdkit -f -v -p 9080 python prototype.py
url="https://[RHEV]/ovirt-engine/api/" username="admin@internal"
disk="6bb55cf2-9d78-41d0-997e-86511db503d4" password=-

Getting the size of the source disk is a little tricky, because it seems
like the exact byte count is not exposed through the ImageIO daemon API.
Get around this with a terrible trick: create an initial transfer ticket
and ask the ImageIO proxy to download the whole thing, but cancel the
transfer and close the HTTP connection as soon as "Content-Length" is
available in the response. That should be an accurate byte count. Then
create a new transfer ticket and connect straight to the ImageIO daemon
to actually read data.